requirements.txt
python-dotenv==1.0.0
datasets==2.14.5
transformers==4.36.0
accelerate==0.22.0
peft==0.5.0
bitsandbytes==0.41.1
evaluate
unidiff==0.7.5
torch
